First compilable draft of package structure
