+++ kompari-cli/README.md
@@ -1,0 +1,3 @@
+# Kompari CLI
+
+Command line interface for Kompari
