+++ kompari-tasks/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "kompari-tasks"
+
+description = "Supportive code for Xtasks based on Kompari"
+keywords = ["image", "report", "diff", "tests"]
+categories = ["graphics", "multimedia::images", "development-tools::testing"]
+readme = "README.md"
+
+
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+repository.workspace = true
+
+[dependencies]
+kompari = { path = "../kompari" }
+kompari-html = { path = "../kompari-html" }
+clap = { workspace = true }
