+++ kompari/README.md
