--- v2/.gitignore
@@ -1,1 +1,0 @@
-/target
