--- v2/Cargo.toml
@@ -1,84 +1,0 @@
-[workspace]
-resolver = "2"
-members = ["kompari", "kompari-app", "kompari-html", "kompari-tasks"]
-
-[workspace.package]
-version = "0.1.0"
-edition = "2021"
-# Keep in sync with RUST_MIN_VER in .github/workflows/ci.yml, with the relevant README.md files,
-# and with the MSRV in the Unreleased section of CHANGELOG.md.
-rust-version = "1.74"
-license = "Apache-2.0 OR MIT"
-repository = "https://github.com/linebender/kompari"
-
-[workspace.dependencies]
-image = { version = "0.25", default-features = false, features = ["png"] }
-
-[workspace.lints]
-# This one may vary depending on the project.
-rust.unsafe_code = "forbid"
-
-# LINEBENDER LINT SET - Cargo.toml - v4
-# See https://linebender.org/wiki/canonical-lints/
-rust.keyword_idents_2024 = "forbid"
-rust.non_ascii_idents = "forbid"
-rust.non_local_definitions = "forbid"
-rust.unsafe_op_in_unsafe_fn = "forbid"
-
-rust.elided_lifetimes_in_paths = "warn"
-rust.let_underscore_drop = "warn"
-rust.missing_debug_implementations = "warn"
-rust.missing_docs = "warn"
-rust.single_use_lifetimes = "warn"
-rust.trivial_numeric_casts = "warn"
-rust.unexpected_cfgs = "warn"
-rust.unit_bindings = "warn"
-rust.unnameable_types = "warn"
-rust.unreachable_pub = "warn"
-rust.unused_import_braces = "warn"
-rust.unused_lifetimes = "warn"
-rust.unused_macro_rules = "warn"
-rust.unused_qualifications = "warn"
-rust.variant_size_differences = "warn"
-
-clippy.too_many_arguments = "allow"
-
-clippy.allow_attributes = "warn"
-clippy.allow_attributes_without_reason = "warn"
-clippy.cast_possible_truncation = "warn"
-clippy.collection_is_never_read = "warn"
-clippy.dbg_macro = "warn"
-clippy.debug_assert_with_mut_call = "warn"
-clippy.doc_markdown = "warn"
-clippy.fn_to_numeric_cast_any = "warn"
-clippy.infinite_loop = "warn"
-clippy.large_include_file = "warn"
-clippy.large_stack_arrays = "warn"
-clippy.match_same_arms = "warn"
-clippy.mismatching_type_param_order = "warn"
-clippy.missing_assert_message = "warn"
-clippy.missing_errors_doc = "warn"
-clippy.missing_fields_in_debug = "warn"
-clippy.missing_panics_doc = "warn"
-clippy.partial_pub_fields = "warn"
-clippy.return_self_not_must_use = "warn"
-clippy.same_functions_in_if_condition = "warn"
-clippy.semicolon_if_nothing_returned = "warn"
-clippy.shadow_unrelated = "warn"
-clippy.should_panic_without_expect = "warn"
-clippy.todo = "warn"
-clippy.unseparated_literal_suffix = "warn"
-clippy.use_self = "warn"
-clippy.wildcard_imports = "warn"
-
-clippy.cargo_common_metadata = "warn"
-clippy.negative_feature_names = "warn"
-clippy.redundant_feature_names = "warn"
-clippy.wildcard_dependencies = "warn"
-# END LINEBENDER LINT SET
-
-
-[profile.ci]
-inherits = "dev"
-[profile.ci.package."*"]
-debug-assertions = true # Keep always on for dependencies for cache reuse.
