--- v2/kompari-app/Cargo.toml
@@ -1,9 +1,0 @@
-[package]
-name = "kompari-app"
-version.workspace = true
-edition.workspace = true
-rust-version.workspace = true
-license.workspace = true
-repository.workspace = true
-
-[dependencies]
