--- v2/kompari-html/Cargo.toml
@@ -1,9 +1,0 @@
-[package]
-name = "kompari-html"
-version.workspace = true
-edition.workspace = true
-rust-version.workspace = true
-license.workspace = true
-repository.workspace = true
-
-[dependencies]
