--- v2/kompari-tasks/Cargo.toml
@@ -1,9 +1,0 @@
-[package]
-name = "kompari-tasks"
-version.workspace = true
-edition.workspace = true
-rust-version.workspace = true
-license.workspace = true
-repository.workspace = true
-
-[dependencies]
